Added a renaming function
Added a new function which ensures all quantified variables have separate names, as well as free variables getting to keep their original names. Not extensively tested, but it seems to work as intended. It seems quite verbose, which could possibly be fixed with functors or the like, but I just wanted to get something working.